テストビルドが通らなくなっていた問題を修正
