+++ Keyboard/Display/KeyboardViewController.swift
@@ -9,8 +9,6 @@
 import SwiftUI
 import UIKit
 
-extension Candidate: ResultViewItemData {}
-
 final private class KeyboardHostingController<Content: View>: UIHostingController<Content> {
     override var preferredScreenEdgesDeferringSystemGestures: UIRectEdge {
         .bottom
